Make separate MSK file
