+++ src/sdr/_modulation/__init__.py
@@ -2,6 +2,7 @@
 A subpackage for various modulation schemes.
 """
 from ._linear import *
+from ._msk import *
 from ._psk import *
 from ._pulse_shapes import *
 from ._symbol_encoding import *
